voxels now have different textures based on value

+++ Examples/NoisePlane/Voxel/Voxel.hpp
@@ -7,7 +7,7 @@
 
 struct Voxel
 {
-	float	Value;
+	unsigned int	Value;
 
 	~Voxel();
 	Voxel();
